LayoutTests/http/tests/referrer-policy/origin/same-origin-expected.txt

Tests the behavior of origin referrer policy when same origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "http://127.0.0.1:8000/"
PASS successfullyParsed is true

TEST COMPLETE
